requirements.txt
discodo==2.0.5
discord.py==1.4.2
tortoise-orm==0.16.16